requirements.txt
# Discord
discord.py>=2.3.0

# LLM
openai>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# API
fastapi>=0.111.0
uvicorn>=0.29.0


